+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 var client = new Client(ctx);
 var httpClient = client.HttpClient;
 ```
-
+* Poll transaction asynchronous with x% overhead of the time the transaction has been running so far.
 ## v0.9.4-alpha
 * Refactored sdk code back into a single package.